+++ Source/ACE.Database/Models/Shard/BiotaPropertiesTitleBook.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace ACE.Database.Models.Shard
+{
+    public partial class BiotaPropertiesTitleBook
+    {
+        public uint Id { get; set; }
+        public uint ObjectId { get; set; }
+        public uint TitleId { get; set; }
+
+        public Biota Object { get; set; }
+    }
+}