+++ Source/ACE.Database/ShardDatabase.cs
@@ -298,12 +298,14 @@ namespace ACE.Database
                 .Include(r => r.BiotaPropertiesBool)
                 .Include(r => r.BiotaPropertiesContract)
                 .Include(r => r.BiotaPropertiesCreateList)
-                .Include(r => r.BiotaPropertiesDID)
+                .Include(r => r.BiotaPropertiesDID)                
                 .Include(r => r.BiotaPropertiesEmote).ThenInclude(emote => emote.BiotaPropertiesEmoteAction)
-                .Include(r => r.BiotaPropertiesEmoteAction)
+                // .Include(r => r.BiotaPropertiesEmoteAction)
+                .Include(r => r.BiotaPropertiesEnchantmentRegistry)
                 .Include(r => r.BiotaPropertiesEventFilter)
+                .Include(r => r.BiotaPropertiesFillCompBook)
                 .Include(r => r.BiotaPropertiesFloat)
-                .Include(r => r.BiotaPropertiesFriendListFriend)
+                // .Include(r => r.BiotaPropertiesFriendListFriend)
                 .Include(r => r.BiotaPropertiesFriendListObject)
                 .Include(r => r.BiotaPropertiesGenerator)
                 .Include(r => r.BiotaPropertiesIID)
@@ -317,6 +319,7 @@ namespace ACE.Database
                 .Include(r => r.BiotaPropertiesSpellBook)
                 .Include(r => r.BiotaPropertiesString)
                 .Include(r => r.BiotaPropertiesTextureMap)
+                .Include(r => r.BiotaPropertiesTitleBook)
                 .FirstOrDefault(r => r.Id == id);
         }
 
@@ -342,6 +345,8 @@ namespace ACE.Database
                     //.Include(r => r.BiotaPropertiesEmote).ThenInclude(emote => emote.BiotaPropertiesEmoteAction)
                     //.Include(r => r.BiotaPropertiesEmoteAction)
                     //.Include(r => r.BiotaPropertiesEventFilter)
+                    .Include(r => r.BiotaPropertiesEnchantmentRegistry)
+                    .Include(r => r.BiotaPropertiesFillCompBook)
                     .Include(r => r.BiotaPropertiesFloat)
                     //.Include(r => r.BiotaPropertiesFriendListFriend)
                     .Include(r => r.BiotaPropertiesFriendListObject)
@@ -356,6 +361,7 @@ namespace ACE.Database
                     .Include(r => r.BiotaPropertiesSpellBar)
                     .Include(r => r.BiotaPropertiesSpellBook)
                     .Include(r => r.BiotaPropertiesString)
+                    .Include(r => r.BiotaPropertiesTitleBook)
                     //.Include(r => r.BiotaPropertiesTextureMap)
                     .FirstOrDefault(r => r.Id == id);
             }