+++ Source/ACE.Server/Command/Handlers/DebugCommands.cs
@@ -941,5 +941,27 @@ namespace ACE.Server.Command.Handlers
             Console.WriteLine("");
             Console.WriteLine($"{string.Join("\n", compatibleCBs.ToArray())}");
         }
+
+        /// <summary>
+        /// Add a specific title to yourself
+        /// </summary
+        [CommandHandler("addtitle", AccessLevel.Developer, CommandHandlerFlag.RequiresWorld, 1,
+            "Add title to yourself", "[titleid]")]
+        public static void HandleAddTitle(Session session, params string[] parameters)
+        {
+            if (uint.TryParse(parameters[0], out var titleId))
+                session.Player.AddTitle(titleId);
+        }
+
+        /// <summary>
+        /// Add all titles to yourself
+        /// </summary
+        [CommandHandler("addalltitles", AccessLevel.Developer, CommandHandlerFlag.RequiresWorld,
+            "Add all titles to yourself")]
+        public static void HandleAddAllTitles(Session session, params string[] parameters)
+        {
+            foreach(CharacterTitle title in Enum.GetValues(typeof(CharacterTitle)))
+                session.Player.AddTitle((uint)title);
+        }
     }
 }