+++ Source/ACE.Server/Network/GameEvent/Events/GameEventCommunicationTransientString.cs
@@ -1,0 +1,11 @@
+namespace ACE.Server.Network.GameEvent.Events
+{
+    public class GameEventCommunicationTransientString : GameEventMessage
+    {
+        public GameEventCommunicationTransientString(Session session, string message)
+            : base(GameEventType.CommunicationTransientString, GameMessageGroup.UIQueue, session)
+        {
+            Writer.WriteString16L(message);
+        }
+    }
+}