+++ changelog.md
@@ -1,5 +1,11 @@
 # ACEmulator Change Log
 
+### 2018-03-13
+[Ripley]
+* Add three new tables to Shard Database.
+* Add CharacterTitle enum.
+* Add proper title support and storage for players. Debug command of `addtitle` and `addalltitles` for testing.
+
 ### 2018-03-11
 
 **ACE-World-16PY world db release v0.0.6+ required with this update**